pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "milly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "markdown>=3.10",
    "pydantic>=2.12.3",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "pytest-mock>=3.15.1",
    "redis>=7.0.1",
    "retry>=0.9.2",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.7",
    "factory-boy>=3.3.3",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.3",
    "types-PyMySQL>=1.1.0",
    "types-retry>=0.9.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
exclude = ["migrations"]
disable_error_code = ["var-annotated", "name-defined"]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "migrations",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B904",  # raise ... from err (too strict for now)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
